+++ packages/form-devtools/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/form-devtools
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [[`c2ecf5d`](https://github.com/TanStack/form/commit/c2ecf5d6df0034d2db982f9b55aed963d94a76a3)]:
+  - @tanstack/form-core@1.27.3
+
+## 0.2.5
+
+### Patch Changes
+
+- Update @tanstack/devtools-utils to 0.0.9, fixes react 17 conflict ([#1892](https://github.com/TanStack/form/pull/1892))
+
+- Updated dependencies []:
+  - @tanstack/form-core@1.27.2
+
 ## 0.2.4
 
 ### Patch Changes